apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
itsdangerous==2.2.0
psycopg[binary]==3.2.9
alembic==1.13.3
pytest==8.3.4
pytest-asyncio==0.25.2
PyJWT==2.10.1
